YaoHIR

YaoHIR.AdjointOperation

YaoHIR.GenericRoutine

YaoHIR.IntrinsicRoutine

YaoHIR.Operation

YaoHIR.Routine

YaoHIR.AdjointOperation — Type

AdjointOperation{P} <: Routine

An AdjointOperation is the adjoint of another Routine.

YaoHIR.GenericRoutine — Type

GenericRoutine{name} <: Function

A GenericRoutine cannot be directly execute on a quantum device. It is a Julia Function that returns Operation, and Operation can be executed on quanutm device.

Note

An instance of GenericRoutine should be treated like Function.

YaoHIR.IntrinsicRoutine — Type

abstract type IntrinsicRoutine <: Routine end

IntrinsicRoutine are the routine that can be executed by the compile target that is pre-defined in the compiler.

YaoHIR.Operation — Type

Operation{P, Args} <: Routine

An Operation is a user defined composite routine that can be called in other Operation or execute on target device.

YaoHIR.Routine — Type

abstract type Routine end

This defines operations that one can execute on a quantum device.